objectModel/Python/requirements.txt
adal==1.2.2
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
idna==2.8
importlib-metadata==0.19
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.720
mypy-extensions==0.4.1
packaging==19.1
pathlib2==2.3.4
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pytest==5.0.1
python-dateutil==2.8.0
regex==2019.8.19
requests==2.22.0
rope==0.14.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.2
